Description:     (Larry & Quincey)     Fix Coverity issues #197, 198 & 199: memory not being released.  (Also clean up other resource leaks in nearby and/or similar code).
Tested on:
    Mac OS X/32 10.6.2 (amazon) debug & production

+++ branches/hdf5_1_8_4_coverity/test/btree2.c
@@ -2735,7 +2735,7 @@ test_insert_lots(hid_t fapl)
 curr_time=1109170019;
 HDfprintf(stderr,"curr_time=%lu\n",(unsigned long)curr_time);
 #endif /* QAK */
-    HDsrandom((unsigned long)curr_time);
+    HDsrandom((unsigned)curr_time);
 
     /*
      * Test inserting many records into v2 B-tree
@@ -6434,7 +6434,7 @@ test_remove_lots(const char *env_h5_drvr, hid_t fapl)
 curr_time = 1163537969;
 HDfprintf(stderr, "curr_time = %lu\n", (unsigned long)curr_time);
 #endif /* QAK */
-    HDsrandom((unsigned long)curr_time);
+    HDsrandom((unsigned)curr_time);
 
     /*
      * Test removing many records into v2 B-tree
